refactor: 태그 알림 전송 기능을 ExerciseService로 이동

+++ src/main/java/com/score/backend/domain/user/UserService.java
@@ -119,6 +119,6 @@ public class UserService {
         if (userOption.isPresent()) {
             return userOption.get().getId();
         }
-        return (long) -1;
+        return -1L;
     }
 }
